Made the definition of u8, u16, etc. cleaner: - we include <asm/types.h> typedefing __u8, __u16 - we #define u8 as __u8
I don't like using #define for this purpose, but this solves many conflicts
with other external programs (waba for instance).

The really clean way would be to use posix definitions like int8_t, but not
every libc has stdint.h, and we	already use u8, u16, etc. almost everywhere.

+++ pgserver/include/pgserver/video.h
@@ -1,4 +1,4 @@
-/* $Id: video.h,v 1.69 2002/01/16 03:56:56 micahjd Exp $
+/* $Id: video.h,v 1.70 2002/01/16 15:25:21 bauermeister Exp $
  *
  * video.h - Defines an API for writing PicoGUI video
  *           drivers
@@ -506,7 +506,7 @@ extern unsigned char trigtab[];
   want to change the mode use vid->setmode
 */
 g_error load_vidlib(g_error (*regfunc)(struct vidlib *v),
-		    s16 xres,s16 yres,s16 bpp,unsigned long flags);
+		    s16 xres,s16 yres,s16 bpp,u32 flags);
 
 /* List of installed video drivers */
 struct vidinfo {
